+++ src/main/java/com/perso/T4C/npc/MarshalTorrhen.java
@@ -9,7 +9,7 @@ import com.perso.T4C.player.BodyPart;
 import com.perso.T4C.spawn.Spawn;
 import java.util.List;
 
-@Spawn(type = "MarshalTorrhen", x = 1700, y = 1560, z = 0, stationary = false, aggressive = false)
+@Spawn(type = "MarshalTorrhen", x = 2430, y = 2440, z = 0, stationary = false, aggressive = false)
 public final class MarshalTorrhen extends ScriptedNpc {
   public static final String SOUND_ATTACK = "Whooshh 1.wav";
   public static final String SOUND_DEATH = "Male Dying 1.wav";